Support flask blueprints and increase code compatibility (#2)
Larger flask projects can be modularized, in the past these has been achieved by using flask modules, but they have been deprecated and were replaced by blueprints. The original flask-genshi implementation worked with both, since blueprints were internally present in the app.modules property. With this property being deprecated the logic needed to be swapped out to support blueprints.

Older versions of flask-genshi had hard-coded 'templates' as directory for finding templates. Flask introduced the 'template_folder' parameter for both apps and blueprints. Initially the logic was shifted to make use of these, but here we add 'templates' as fall-back so behavior is consistent with earlier versions if the parameter is unset.

The original author added jinja filters as top-level access somewhere between version 0.5.1 and the latest master on their repository. This caused a naming conflict for builtins commands that have a matching jina filter like lower, with the jinja filter overshadowing the builtin. This caused problems especially for functions with different signature between the two like max. The jinja filter max only accepts a single iterable, while Python built-in max supports a variable number of inputs to iterate over.